protocol change to host-callback design
This design fixes design issues where reconnections are dependent on the
host establishing new idle connections, and rapid client
disconnect/reconnect could easily fail.

The new design has the clients wait on an idle nexus until the host
calls back on a different websocket. This allows the host to decide at
runtime which connections to accept, and allows multiple clients to
connect to the same service without risk of confusing the sockets.