tcp-forward example: fix service-name and port handling
This code reflects an older way of handling multiple services, where the
port number was encoded in the service name. Now the name->port mapping
is handled at runtime by the host.

The tcp-forward example has has CLI parameters for the service name and
target port. It will only accept connections using that exact service
name and always connect to the target host and port specified in the CLI
arguments.