enable keepalive messages on the host control connection
The host connection is idle for long periods, and might fall victim to
bits of network infrastructure that will drop a TCP FIN packet. This
would mean that the socket has been closed by the server but the host
doesn't know that and will never attempt to reconnect.

To prevent this, add keepalive messages every 120 seconds. This has two
effects:
1. It may make e.g. NAT routers less likely to silently forget open
connections
2. A host with a mysteriously dead connection will discover it's dead
when it attempts to send and gets an error.

Cloudflare workers infrastructure auto-replies to a websocket ping with
a pong message, so this extra traffic won't cause a hibernating worker
to wake.