portal-router: terminate data connection if forwarding fails
There's no point holding a data connection open if we can't find the
peer or can't send to it. Even if whatever's wrong resolves itself, we
would have dropped some of the data.

Also terminate the connection if an improper text message is received.
Only the server is allowed to send those.