portal-router: close stale host control sockets
If a host connects but we have a previous host control socket, close it.
This may improve situations where the previous socket appears
unresponsive, and can take a while to die off on its own.

Add a code and reason message to the close, and log them in the
host code. This will help if a host gets cloned and are caught it a loop
bumping each other off the portal.